Dry controllers with concerns (closes #6)
