+++ app/controllers/siege_machines_controller.rb
@@ -1,4 +1,7 @@
 class SiegeMachinesController < ApplicationController
+  include OperationResultToHash
+  include SetModelFromIds
+
   before_action :authorize_access_request!
   before_action :set_camp, only: %i[index]
   before_action :set_character, only: %i[arm]
@@ -31,37 +34,7 @@ class SiegeMachinesController < ApplicationController
 
   private
 
-  def set_camp
-    @camp = Camp.find params[:camp_id]
-  rescue ActiveRecord::RecordNotFound
-    render json: { error: 'camp not found' }, status: :not_found
-  end
-
-  def set_character
-    @character = Character.find params[:character_id]
-  rescue ActiveRecord::RecordNotFound
-    render json: { error: 'character not found' }, status: :not_found
-  end
-
-  def set_siege_machine
-    @siege_machine = SiegeMachine.find params[:id]
-  rescue ActiveRecord::RecordNotFound
-    render json: { error: 'siege machine not found' }, status: :not_found
-  end
-
   def arming
     @arming ||= SiegeMachine::Arm.call(siege_machine: @siege_machine, character: @character)
-  end
-
-  def camp_hash
-    CampBlueprint.render_as_hash @operation_result[:camp]
-  end
-
-  def castle_hash
-    CastleBlueprint.render_as_hash @operation_result[:castle]
-  end
-
-  def weapon_hash
-    SiegeMachineBlueprint.render_as_hash @operation_result[:siege_machine]
   end
 end
